==> Rmd/1_Preprocess_app.Rmd <==
---
title: "Preprocess_app"
author: "kotsubo takuto"
output: 
    html_document:
      md_extensions: -ascii_identifiers
      toc: true
      toc_depth: 3
---

# About application data

## Data description

- application_{train|test}.csv (122|121 variables)
    - Key variable: SK_ID_CURR
    - Objective variable: TARGET(1:bad, 0: good)
    - About: all application data

## Task

# Setting{.tabset .tabset-fade .tabset-pills}

## knitr option

```{r reset, include=FALSE}
# 初期化
rm(list = ls())
```

```{r set up, message=FALSE}
# set directory
setwd("~/Desktop/Home_Credit_Kaggle/") 
# max.print 
options(max.print="200", digits=5)
# Global options
library(knitr)
opts_chunk$set(echo=TRUE,
	             prompt=FALSE,
               tidy=TRUE,
               comment=NA,
               message=FALSE,
               warning=FALSE)
opts_knit$set(width=75)
```

## Library package

```{r package, message=FALSE}
library(tidyverse)
library(readr) # for csv
library(summarytools) # summary easily for EDA
library(GGally) # ggpairs
library(skimr) 
library(janitor)
library(missForest)
library(imputeMissings)
library(mi)
library(doParallel)
library(qgraph)
library(psych)
library(fastcluster)
library(Rtsne)
```

## Load funciton

```{r cache=FALSE}
source('~/Desktop/Home_Credit_Kaggle/script/function.R') # for preprocessing
source('~/Desktop/Home_Credit_Kaggle/script/makedummies.R') # for preprocessing
```

## read csv

- Define missing values: na = c("XNA","NA","","NaN","?",...)
- Convert charcter's columns into factor's columns 

```{r}
# train data and test data
app_train <- read_csv("~/Desktop/Home_Credit_Kaggle/input/csv_imp0/application_train.csv",na = c("XNA","NA","","NaN","?")) %>% 
  mutate_if(is.character, funs(factor(.))) 
  
app_test <- read_csv("~/Desktop/Home_Credit_Kaggle/input/csv_imp0/application_test.csv",na = c("XNA","NA","","NaN","?")) %>% 
  mutate_if(is.character, funs(factor(.)))
```

# Preprocess

- Convert outliers into NA
- Add `_imp` to the complemented column
- Add `_add` to the made new column

## Check outliers

- Convert outliers into NA

```{r}
app_train <- app_train %>% 
  mutate(days_employed_app = ifelse(days_employed_app == 365243, NA, days_employed_app),
         days_last_phone_change_app = ifelse(days_last_phone_change_app == 0, NA, days_last_phone_change_app))
app_test <- app_test %>% 
  mutate(days_employed_app = ifelse(days_employed_app == 365243, NA, days_employed_app),
         days_last_phone_change_app = ifelse(days_last_phone_change_app == 0, NA, days_last_phone_change_app))
```

## impute missing values{.tabset .tabset-fade .tabset-pills}

- Complement missing values according to feature of data
- Add `_imp` to the complemented column
- Estimate missing values according to algorithms such as multiple imputation method (*later*)

### code_gender_app

- 少ないので放置 (欠損値:4つ)

```{r eval=FALSE}
app_train %>% 
  tabyl(code_gender_app) %>% 
  
app_test %>% 
  tabyl(code_gender_app) %>% 
```

### name_type_suite_app

- 同行者の情報
- 無記載 = 同行者なしと考える

```{r eval=TRUE}
app_train <- app_train %>% 
  mutate(name_type_suite_app_imp = replace(name_type_suite_app, is.na(name_type_suite_app),"Unaccompanied")) 
app_test <- app_test %>% 
  mutate(name_type_suite_app_imp = replace(name_type_suite_app, is.na(name_type_suite_app),"Unaccompanied")) 
```

### occupation_type_app

- other, フリーターの項目がない
- 所属=NA and 職業=NA -> 無職,無所属
- 所属=NA, 職業=NA -> other

```{r eval=FALSE}
app_train %>% tabyl(occupation_type_app) 
app_test %>% tabyl(occupation_type_app)
```

```{r eval=TRUE}
app_train <- app_train %>% 
  mutate(occupation_type_app_imp = ifelse(!is.na(occupation_type_app),occupation_type_app %>% as.character(),
                                      ifelse(is.na(organization_type_app),"Unemployed","other"))) %>% 
  mutate(occupation_type_app_imp = occupation_type_app_imp %>% as.factor())

app_test <- app_test %>% 
  mutate(occupation_type_app_imp = ifelse(!is.na(occupation_type_app),occupation_type_app %>% as.character(),
                                      ifelse(is.na(organization_type_app),"Unemployed","other"))) %>% 
  mutate(occupation_type_app_imp = occupation_type_app_imp %>% as.factor())
```

### organization_type_app 

- 所属=NAはほぼ, 職業=NA (例外は2件のみ)
- 所属=NA and 職業=NA -> 無職,無所属
- 所属=NA, 職業=NA -> other

```{r eval=FALSE}
app_train %>% tabyl(organization_type_app,occupation_type_app)
#app_train %>% select(organization_type_app_imp,occupation_type_app_imp) %>% table(exclude = NULL) 
```

```{r eval=TRUE}
app_train <- app_train %>% 
  mutate(organization_type_app_imp = ifelse(!is.na(organization_type_app),organization_type_app %>% as.character(),
                                      ifelse(is.na(occupation_type_app),"Unemployed","Other"))) %>% 
  mutate(organization_type_app_imp = organization_type_app_imp %>% as.factor())

app_test <- app_test %>% 
  mutate(organization_type_app_imp = ifelse(!is.na(organization_type_app),organization_type_app %>% as.character(),
                                      ifelse(is.na(occupation_type_app),"Unemployed","other"))) %>% 
  mutate(organization_type_app_imp = organization_type_app_imp %>% as.factor())
```

### own_car_age_app

- 車の年数
- 車の所持データとほぼ一致(5件: 車保持かつ年数不明があった)
- 放置

```{r eval=FALSE}
app_train %>% tabyl(flag_own_car_app,own_car_age_app)
```

### cnt_fam_members_app

- 家族の人数
- 欠損値は一人と考える

```{r eval=FALSE}
# 子どもの人数 / 家族の人数
app_train %>% tabyl(cnt_children_app,cnt_fam_members_app) 
```

```{r eval=TRUE}
app_train <- app_train %>% 
  mutate(cnt_fam_members_app_imp = ifelse(is.na(cnt_fam_members_app),1,cnt_fam_members_app )) 
app_test <- app_test %>% 
  mutate(cnt_fam_members_app_imp = ifelse(is.na(cnt_fam_members_app),1,cnt_fam_members_app )) 
```

### ext_source_{1,2,3}_app

- data list
    - ext_source_1
    - ext_source_2
    - ext_source_3

- 何かの外部データ, 影響が強いので要検証
- 収入とクレジットを元に, 多重代入法による欠損値補完(error)

```{r eval=FALSE}
app_train %>% 
  select(ext_source_1_app,ext_source_2_app,ext_source_3_app,target_app) %>% 
  na.omit() %>% 
  ggpairs(cardinality_threshold=20)
```

### ~_avg_app

- data list
    - apartments_avg_app
    - basementarea_avg_app
    - years_beginexpluatation_avg_app
    - years_build_avg_app
    - commonarea_avg_app
    - elevators_avg_app
    - entrances_avg_app
    - floorsmax_avg_app
    - floorsmin_avg_app
    - landarea_avg_app
    - livingapartments_avg_app
    - livingarea_avg_app
    - nonlivingapartments_avg_app
    - nonlivingarea_avg_app

- 住居等に関数正規化された情報
- 平均項目
- 14項目
- 保留

```{r eval=FALSE}
app_train %>% 
  select_if(grepl("_avg_",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

app_train %>% 
  select_if(grepl("_avg_",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)
```

### ~_medi_app

- data list
    - apartments_medi_app
    - basementarea_medi_app
    - years_beginexpluatation_medi_app
    - years_build_medi_app
    - commonarea_medi_app
    - elevators_medi_app
    - entrances_medi_app
    - floorsmax_medi_app
    - floorsmin_medi_app
    - landarea_medi_app
    - livingapartments_medi_app
    - livingarea_medi_app
    - nonlivingapartments_medi_app
    - nonlivingarea_medi_app

- 住居等に関数正規化された情報
- 中央値
- 14項目
- 保留

```{r eval=FALSE}
app_train %>% 
  select_if(grepl("_medi_",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

app_train %>% 
  select_if(grepl("_medi_",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)
```

### ~_mode_app

建物に関する諸情報(標準化済み)

- factor data
    - fondkaprement_mode_app
    - housetype_mode_app
    - wallsmaterial_mode_app
    - totalarea_mode_app
    - emergencystate_mode_app (No or Yes) -> 欠損値 = NOとする
    
- numeric data
    - apartments_mode_app
    - basementarea_mode_app
    - years_beginexpluatation_mode_app
    - years_build_mode_app
    - commonarea_mode_app
    - elevators_mode_app
    - entrances_mode_app
    - floorsmax_mode_app
    - floorsmin_mode_app
    - landarea_mode_app
    - livingapartments_mode_app
    - livingarea_mode_app
    - nonlivingapartments_mode_app
    - nonlivingarea_mode_app

- numeric: 14項目
- factor: 5項目

#### factor value 

```{r eval=FALSE}
app_train %>% tabyl(housetype_mode_app,wallsmaterial_mode_app) 
app_train %>% tabyl(fondkapremont_mode_app,wallsmaterial_mode_app)  
app_train %>% tabyl(housetype_mode_app,fondkapremont_mode_app) 
app_train %>% select(totalarea_mode_app) %>% graphics::hist(breaks = 30)

app_train %>% 
  select(housetype_mode_app,wallsmaterial_mode_app,fondkapremont_mode_app,totalarea_mode_app) %>% 
  na.omit() %>% 
  ggpairs()
```

```{r eval=TRUE}
## impute 1
app_train <- app_train %>% 
  mutate(emergencystate_mode_app_imp = ifelse(is.na(emergencystate_mode_app),"No",emergencystate_mode_app %>% as.character())) %>% 
  mutate(emergencystate_mode_app_imp = emergencystate_mode_app_imp %>% as.factor())
app_test <- app_test %>% 
  mutate(emergencystate_mode_app_imp = ifelse(is.na(emergencystate_mode_app),"No",emergencystate_mode_app %>% as.character())) %>% 
  mutate(emergencystate_mode_app_imp = emergencystate_mode_app_imp %>% as.factor())
```

#### numeric value

```{r eval=FALSE}
# exclude list
tmp <- c("housetype_mode_app","wallsmaterial_mode_app","fondkapremont_mode_app","totalarea_mode_app","emergencystate_mode_app")
app_train %>% 
  select_if(grepl("_mode_",names(.))) %>% 
  select_if(!grepl(paste(tmp, collapse="|"),names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

app_train %>% 
  select_if(grepl("_mode_",names(.))) %>% 
  select_if(!grepl(paste(tmp, collapse="|"),names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)
```

### ~_circle_app

- data list
    - obs_30_cnt_social_circle_app
    - def_30_cnt_social_circle_app
    - obs_60_cnt_social_circle_app
    - def_60_cnt_social_circle_app
- 観測されていない => 0

```{r eval=FALSE}
app_train %>% 
  select_if(grepl("_circle_",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

app_train %>% 
  select_if(grepl("_circle_",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)
```

```{r eval=TRUE}
app_train <- app_train %>% 
  mutate(obs_30_cnt_social_circle_app_imp = ifelse(is.na(obs_30_cnt_social_circle_app),0,obs_30_cnt_social_circle_app )) %>% 
  mutate(def_30_cnt_social_circle_app_imp = ifelse(is.na(def_30_cnt_social_circle_app),0,def_30_cnt_social_circle_app )) %>% 
  mutate(obs_60_cnt_social_circle_app_imp = ifelse(is.na(obs_60_cnt_social_circle_app),0,obs_60_cnt_social_circle_app )) %>% 
  mutate(def_60_cnt_social_circle_app_imp = ifelse(is.na(def_60_cnt_social_circle_app),0,def_60_cnt_social_circle_app )) 
app_test <- app_test %>% 
  mutate(obs_30_cnt_social_circle_app_imp = ifelse(is.na(obs_30_cnt_social_circle_app),0,obs_30_cnt_social_circle_app )) %>% 
  mutate(def_30_cnt_social_circle_app_imp = ifelse(is.na(def_30_cnt_social_circle_app),0,def_30_cnt_social_circle_app )) %>% 
  mutate(obs_60_cnt_social_circle_app_imp = ifelse(is.na(obs_60_cnt_social_circle_app),0,obs_60_cnt_social_circle_app )) %>% 
  mutate(def_60_cnt_social_circle_app_imp = ifelse(is.na(def_60_cnt_social_circle_app),0,def_60_cnt_social_circle_app )) 
```

### days_last_phone_change_app

- 欠損値一人
- 保留

```{r eval=FALSE}
app_train %>% tabyl(days_last_phone_change_app)
```

### amt_~

#### part1 

- data list
    - amt_income_total_app: 総収入
    - amt_credit_app: ローンのクレジット額
    - amt_annuity_app: 年金ローン
    - amt_goods_price_app: 消費者ローン

- 3つのローンの相関関係強い, 欠損値補完に使える！！

```{r eval=FALSE}
app_train %>%
  select(amt_income_total_app,amt_credit_app,amt_annuity_app,amt_goods_price_app)
  na.omit() %>% 
  ggpairs()
```

```{r eval=TRUE}
# 補完に使用するデータ
data1 <- app_train %>% select(amt_credit_app,amt_annuity_app,amt_goods_price_app) 
data2 <- app_test %>% select(amt_credit_app,amt_annuity_app,amt_goods_price_app) 
# 欠損値を含むデータ
patterns <- c("amt_annuity_app","amt_goods_price_app")

# 欠損値補完 (1 iter : 2000 second)
tmp1 <- ImputeMissingValueMI(data1,patterns)
tmp2 <- ImputeMissingValueMI(data2,patterns)
  
# データ結合
app_train <- cbind(app_train,tmp1)
app_test <- cbind(app_test,tmp2)
```

#### part2

- data list
    - amt_req_credit_bureau_hour_app
    - amt_req_credit_bureau_day_app
    - amt_req_credit_bureau_week_app
    - amt_req_credit_bureau_mon_app
    - amt_req_credit_bureau_qrt_app
    - amt_req_credit_bureau_year_app

- 何かの紹介件数 NA-> 0補完
- 関係性低い

```{r eval=FALSE}
app_train %>% 
  select(amt_req_credit_bureau_hour_app, amt_req_credit_bureau_day_app, amt_req_credit_bureau_week_app,
         amt_req_credit_bureau_mon_app, amt_req_credit_bureau_qrt_app, amt_req_credit_bureau_year_app) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

app_train %>% 
  select(amt_req_credit_bureau_hour_app, amt_req_credit_bureau_day_app, amt_req_credit_bureau_week_app,
         amt_req_credit_bureau_mon_app, amt_req_credit_bureau_qrt_app, amt_req_credit_bureau_year_app) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)
```

```{r eval=TRUE}
app_train <- app_train %>% 
  mutate(amt_req_credit_bureau_hour_app_imp = ifelse(is.na(amt_req_credit_bureau_hour_app),0,amt_req_credit_bureau_hour_app)) %>% 
  mutate(amt_req_credit_bureau_day_app_imp = ifelse(is.na(amt_req_credit_bureau_day_app),0,amt_req_credit_bureau_day_app)) %>% 
  mutate(amt_req_credit_bureau_week_app_imp = ifelse(is.na(amt_req_credit_bureau_week_app),0,amt_req_credit_bureau_week_app)) %>% 
  mutate(amt_req_credit_bureau_mon_app_imp = ifelse(is.na(amt_req_credit_bureau_mon_app),0,amt_req_credit_bureau_mon_app)) %>% 
  mutate(amt_req_credit_bureau_qrt_app_imp = ifelse(is.na(amt_req_credit_bureau_qrt_app),0,amt_req_credit_bureau_qrt_app)) %>% 
  mutate(amt_req_credit_bureau_year_app_imp = ifelse(is.na(amt_req_credit_bureau_year_app),0,amt_req_credit_bureau_year_app)) 

app_test <- app_test %>% 
  mutate(amt_req_credit_bureau_hour_app_imp = ifelse(is.na(amt_req_credit_bureau_hour_app),0,amt_req_credit_bureau_hour_app)) %>% 
  mutate(amt_req_credit_bureau_day_app_imp = ifelse(is.na(amt_req_credit_bureau_day_app),0,amt_req_credit_bureau_day_app)) %>% 
  mutate(amt_req_credit_bureau_week_app_imp = ifelse(is.na(amt_req_credit_bureau_week_app),0,amt_req_credit_bureau_week_app)) %>% 
  mutate(amt_req_credit_bureau_mon_app_imp = ifelse(is.na(amt_req_credit_bureau_mon_app),0,amt_req_credit_bureau_mon_app)) %>% 
  mutate(amt_req_credit_bureau_qrt_app_imp = ifelse(is.na(amt_req_credit_bureau_qrt_app),0,amt_req_credit_bureau_qrt_app)) %>% 
  mutate(amt_req_credit_bureau_year_app_imp = ifelse(is.na(amt_req_credit_bureau_year_app),0,amt_req_credit_bureau_year_app)) 
```

## add new column{.tabset .tabset-fade .tabset-pills}

- Add `_add` to the made new column

### everything

```{r}
app_train <- app_train %>%
  mutate(credit_annuity_perc_add = amt_credit_app/amt_annuity_app,
         credit_goods_perc_add = amt_credit_app/amt_goods_price_app,
         credit_income_perc_add = amt_credit_app / amt_income_total_app,
         income_children_perc_add = amt_income_total_app / (1 + cnt_children_app),
         empoloyed_birth_perc_add = days_employed_app / days_birth_app)
app_test <- app_test %>%          
  mutate(credit_annuity_perc_add = amt_credit_app/amt_annuity_app,
         credit_goods_perc_add = amt_credit_app/amt_goods_price_app,
         credit_income_perc_add = amt_credit_app / amt_income_total_app,
         income_children_perc_add = amt_income_total_app / (1 + cnt_children_app),
         empoloyed_birth_perc_add = days_employed_app / days_birth_app)
          
         # OWN_CAR_BIRTH_PERC_add = OWN_CAR_AGE / DAYS_BIRTH,
         # CHILDREN_MEMBERS_PERC_add = CNT_CHILDREN / cnt_fam_members_app,
         # OWN_CAR_EMPLOY_PERC_add = OWN_CAR_AGE / DAYS_EMPLOYED,
         # INCOME_MEMBERS_PERC_add = AMT_INCOME_TOTAL / cnt_fam_members_app,
         # PHONE_TO_BIRTH_PERC_add = DAYS_LAST_PHONE_CHANGE / DAYS_BIRTH,
         # PHONE_TO_EMPLOY_PERC_add = DAYS_LAST_PHONE_CHANGE / DAYS_EMPLOYED,
         # CREDIT_CHILDEREN_PERC_add = AMT_CREDIT / (1 + CNT_CHILDREN),
         # CREDIT_MEMBERS_PERC_add = AMT_CREDIT / cnt_fam_members_app,
         # CNT_NON_CHILD_add = cnt_fam_members_app - CNT_CHILDREN,
         # INCOME_NO_CHILD_PERC_add = AMT_INCOME_TOTAL / CNT_NON_CHILD_add,
         # CREDIT_NO_CHILD_PERC_add = AMT_CREDIT / CNT_NON_CHILD_add 
```

### ext_source_{1,2,3}_app

- data list
    - ext_source_1_app
    - ext_source_2_app
    - ext_source_3_app

- 全てNAのものは172個
- 合計値と乗算値

```{r eval=FALSE}
app_train %>% 
  filter(ext_source_1_app %>% is.na()) %>% 
  filter(ext_source_2_app %>% is.na()) %>% 
  filter(ext_source_3_app %>% is.na()) %>% 
  tabyl(target_app)
```

```{r eval=TRUE}
app_train <- app_train %>% 
  mutate(tmp1 = ext_source_1_app,tmp2 = ext_source_2_app,tmp3 = ext_source_3_app) %>% 
  mutate_at(vars(matches("tmp",.)), funs(ifelse(is.na(.),0,.))) %>% 
  mutate(ext_source_sum_add = round(tmp1 + tmp2 + tmp3, digits = 4)) %>% 
  mutate(ext_source_multi_add = round(tmp1 * tmp2 * tmp3, digits = 4)) %>%
  mutate(ext_source_mean_add = round(mean(c(tmp1,tmp2,tmp3),na.rm=TRUE),digits = 4)) %>% 
  select(-c(tmp1,tmp2,tmp3))

app_test <- app_test %>% 
  mutate(tmp1 = ext_source_1_app,tmp2 = ext_source_2_app,tmp3 = ext_source_3_app) %>% 
  mutate_at(vars(matches("tmp",.)), funs(ifelse(is.na(.),0,.))) %>% 
  mutate(ext_source_sum_add = round(tmp1 + tmp2 + tmp3, digits = 4)) %>% 
  mutate(ext_source_multi_add = round(tmp1 * tmp2 * tmp3, digits = 4)) %>%
  mutate(ext_source_mean_add = round(mean(c(tmp1,tmp2,tmp3),na.rm=TRUE),digits = 4)) %>% 
  select(-c(tmp1,tmp2,tmp3))
```

```{r eval=FALSE}
patterns <- c("ext_source_tsne_1_app","ext_source_tsne_2_app")
# make new column
tmp1 <- app_train %>% 
  select_if(grepl("ext_source",names(.))) %>% 
  mutate_all(funs(ifelse(is.na(.),0,.))) %>% 
  ImputeMissingValueTSNE(patterns)
tmp2 <- app_test %>% 
  select_if(grepl("ext_source",names(.))) %>% 
  mutate_all(funs(ifelse(is.na(.),0,.))) %>% 
  ImputeMissingValueTSNE(patterns)
# combine data
app_train <- cbind(app_train,tmp1)
app_test <- cbind(app_test,tmp2)
```

### avg

- 欠損値に0を代入してRtsneを実行

```{r}
patterns <- c("avg_tsne_1_app", "avg_tsne_2_app")
# make new column
tmp1 <- app_train %>% 
  select_if(grepl("_avg_",names(.))) %>% 
  mutate_all(funs(ifelse(is.na(.),0,.))) %>% 
  ImputeMissingValueTSNE(patterns)
tmp2 <- app_test %>% 
  select_if(grepl("_avg_",names(.))) %>% 
  mutate_all(funs(ifelse(is.na(.),0,.))) %>% 
  ImputeMissingValueTSNE(patterns)
# combine data
app_train <- cbind(app_train,tmp1)
app_test <- cbind(app_test,tmp2)
```

### mode

- 欠損値に0を代入してRtsneを実行

```{r}
# exclude
tmp <- c("housetype_mode_app","wallsmaterial_mode_app","fondkapremont_mode_app","totalarea_mode_app","emergencystate_mode_app")

# list
patterns <- c( "mode_tsne_1_app", "mode_tsne_2_app")

tmp1 <- app_train %>% 
  select_if(grepl("_mode",names(.))) %>% 
  select_if(!grepl(paste(tmp, collapse="|"),names(.))) %>% 
  mutate_all(funs(ifelse(is.na(.),0,.))) %>% 
  ImputeMissingValueTSNE(patterns)

tmp2 <- app_test %>% 
  select_if(grepl("_mode",names(.))) %>% 
  select_if(!grepl(paste(tmp, collapse="|"),names(.))) %>% 
  mutate_all(funs(ifelse(is.na(.),0,.))) %>% 
  ImputeMissingValueTSNE(patterns)

# データ結合
app_train <- cbind(app_train,tmp1)
app_test <- cbind(app_test,tmp2)
```

### medi

- 欠損値に0を代入してRtsneを実行

```{r}
# list
patterns <- c("medi_tsne_1_app", "medi_tsne_2_app")

# data
tmp1 <- app_train %>% 
  select_if(grepl("_medi_",names(.))) %>% 
  mutate_all(funs(ifelse(is.na(.),0,.))) %>% 
  ImputeMissingValueTSNE(patterns)

tmp2 <- app_test %>%
  select_if(grepl("_medi_",names(.))) %>% 
  mutate_all(funs(ifelse(is.na(.),0,.))) %>% 
  ImputeMissingValueTSNE(patterns)

# データ結合
app_train <- cbind(app_train,tmp1)
app_test <- cbind(app_test,tmp2)
```

## output for csv

- Save the csv files in the `input/csv_imp1`
- Convert charcter's columns into factor's columns 

```{r eval=TRUE}
app_train %>% 
  mutate_if(is.factor, funs(as.numeric)) %>% 
  write_csv("~/Desktop/Home_Credit_Kaggle/input/csv_imp1/application_train_imp.csv")
app_test %>% 
  mutate_if(is.factor, funs(as.numeric)) %>% 
  write_csv("~/Desktop/Home_Credit_Kaggle/input/csv_imp1/application_test_imp.csv")
```
