==> Rmd/1_Preprocess_bureau.Rmd <==
---
title: "Preprocess_bureau"
author: "kotsubo takuto"
output: 
    html_document:
      md_extensions: -ascii_identifiers
      toc: true
      toc_depth: 3
---

## Data description

- bureau.csv (17 variables)
    - Key variable: SK_ID_CURR, SK_ID_BUREAU 
    - About: 以前使用していた, 融資の履歴
    - Credit Bureau: 信用調査所
- bureau_balance.csv (3 variables)
    - Key variable: SK_ID_BUREAU
    - About: `bureau.csv`の補助データ
    - MONTHS_BALANCE: 利用日に対応した, 月次のバランス?
    - STATUS: 月次のステータス, (active, closed, DPD0-30,… [C means closed, X means status unknown, 0 means no DPD, 1 means maximal did during month between 1-30, 2 means DPD 31-60,
    3 means DPD 61-90, 4 means DPD 91-120, 5 means DPD 120+ or sold or written off ] )

# Setting{.tabset .tabset-fade .tabset-pills}

## knitr option

```{r reset, include=FALSE}
# 初期化
rm(list = ls())
```

```{r set up, message=FALSE}
# set directory
setwd("~/Desktop/Home_Credit_Kaggle/") 
# max.print 
options(max.print="200", digits=5)
# Global options
library(knitr)
opts_chunk$set(echo=TRUE,
               cache = TRUE,
	             prompt=FALSE,
               tidy=TRUE,
               comment=NA,
               message=FALSE,
               warning=FALSE)
opts_knit$set(width=75)
```

## Library package

- 必要なパッケージを適宜追加

```{r package, message=FALSE}
library(tidyverse)
library(readr) # for csv
library(summarytools) # summary easily for EDA
library(GGally) # ggpairs
library(skimr) 
library(missForest)
library(imputeMissings)
library(mi)
library(doParallel)
library(qgraph)
library(psych)
library(fastcluster)
```

## Load funciton

```{r cache=FALSE}
source('~/Desktop/Home_Credit_Kaggle/script/function.R') # for preprocessing
source('~/Desktop/Home_Credit_Kaggle/script/makedummies.R') # for preprocessing
```

## read csv

- Missing value
na = c("XNA","NA","","NaN","?"): 以下の設定で naを計上する.

- convert charcter type into factor type

```{r}
# クレジット利用履歴 (by Credit Bureau)
bureau <- read_csv("~/Desktop/Home_Credit_Kaggle/csv/bureau.csv",na = c("XNA","NA","","NaN","?")) %>% 
  mutate_if(is.character, funs(factor(.)))
bureau_balance <- read_csv("~/Desktop/Home_Credit_Kaggle/csv/bureau_balance.csv",na = c("XNA","NA","","NaN","?")) %>% 
  mutate_if(is.character, funs(factor(.)))
```

# Preprocess

- Check for imputation values
- Describe the details on the explanation sheet
- Add `_imp` to the supplemented column

## impute missing values{.tabset .tabset-fade .tabset-pills}

- 検証内容
    - データ特性に従って欠損値を推定
    - 多重代入法などのアルゴリムに従って, 欠損値を推定

### DAYS_CREDIT_ENDDATE 

```{r eval=FALSE}
bureau %>%
  select(DAYS_CREDIT) %>% 
  hist()
bureau %>% 
  filter(is.na(DAYS_CREDIT_ENDDATE)) %>% 
  select(DAYS_CREDIT) %>% 
  hist()

bureau %>% 
  select(CREDIT_DAY_OVERDUE) %>% 
  table()
bureau %>% 
  filter(is.na(DAYS_CREDIT_ENDDATE)) %>% 
  select(CREDIT_DAY_OVERDUE) %>% 
  table()
```

### DAYS_ENDDATE_FACT 

- 欠損値の中で内容が違う気がする

```{r eval=FALSE}
bureau %>% 
  filter(!is.na(DAYS_ENDDATE_FACT)) %>% 
  select(CREDIT_ACTIVE) %>% 
  table()

bureau %>% 
  filter(is.na(DAYS_ENDDATE_FACT)) %>% 
  select(CREDIT_ACTIVE) %>% 
  table()
```

### DAYS_???

- DAYS_CREDIT
- CREDIT_DAY_OVERDUE
- DAYS_CREDIT_ENDDATE
- DAYS_ENDDATE_FACT
- DAYS_CREDIT_UPDATE

```{r eval=FALSE}
bureau %>% 
  select_if(grepl("DAYS_",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

bureau %>% 
  select_if(grepl("DAY",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)

bureau %>% 
  select_if(grepl("DAY",names(.))) %>% 
  na.omit() %>% 
  ggpairs(cardinality_threshold=20)
```

```{r}
# data
data1 <- bureau %>% select_if(grepl("DAY",names(.)))
patterns <- c("DAYS_CREDIT_ENDDATE","DAYS_ENDDATE_FACT")
# 欠損値補完 
tmp1 <- ImputeMissingValueMI(data1,patterns)
# データ結合
bureau <- cbind(bureau,tmp1)
```

### AMT_???

- AMT_CREDIT_MAX_OVERDUE
- AMT_CREDIT_SUM_OVERDUE
- AMT_CREDIT_SUM
- AMT_CREDIT_SUM_DEBT 
- AMT_CREDIT_SUM_LIMIT 
- AMT_ANNUITY

```{r eval=FALSE}
bureau %>% 
  select_if(grepl("AMT_",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

bureau %>% 
  select_if(grepl("AMT_",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)
```

```{r}
# data
data1 <- bureau %>% select_if(grepl("AMT",names(.)))
patterns <- c("AMT_CREDIT_MAX_OVERDUE","AMT_CREDIT_SUM","AMT_CREDIT_SUM_DEBT","AMT_CREDIT_SUM_LIMIT","AMT_ANNUITY")
# 欠損値補完 
tmp1 <- ImputeMissingValueMI(data1,patterns)
# データ結合
bureau <- cbind(bureau,tmp1)
```

## add new column{.tabset .tabset-fade .tabset-pills}

- 新しい特徴量を作成する

### bureau_balance

- 自分で考えてみたが, 一旦関数を使った統一的処理

```{r eval=FALSE}
tmp1 <- bureau_balance %>% 
  filter(STATUS == "C") %>% 
  group_by(SK_ID_BUREAU) %>%
  summarise(STATUS_C_add = n())

tmp2 <- bureau_balance %>% 
  filter(STATUS == "X") %>% 
  group_by(SK_ID_BUREAU) %>%
  summarise(STATUS_X_add = n())

tmp3 <- bureau_balance %>% 
  mutate_if(is.factor, funs(as.numeric(.))) %>% 
  filter(STATUS < 7) %>% 
  mutate(STATUS = STATUS - 1) %>% 
  group_by(SK_ID_BUREAU) %>% 
  summarise(STATUS_MEAN_add = mean(STATUS))

tmp4 <- bureau_balance %>% 
  group_by(SK_ID_BUREAU) %>%
  summarise(MONTHS_BALANCE_MIN_add = min(MONTHS_BALANCE),
            MONTHS_BALANCE_MAX_add = max(MONTHS_BALANCE))

# combine
bureau <- bureau %>% 
  left_join(tmp1,by = "SK_ID_BUREAU") %>% 
  left_join(tmp2,by = "SK_ID_BUREAU") %>% 
  left_join(tmp3,by = "SK_ID_BUREAU") %>% 
  left_join(tmp4,by = "SK_ID_BUREAU") 
# remove
rm(tmp1,tmp2,tmp3,tmp4)
```

## Summarize

- *NAがInf or - Infになることに注意する*(max,min 計算時)

```{r}
bureau <- makedummies(bureau_balance,basal_level = TRUE) %>% # make dummy
  SummarizeFunc(group = "SK_ID_BUREAU") %>%                  # summarize
  select(-c(mean,sum,sd)) %>%                                # not need remove
  left_join(bureau,.,by = "SK_ID_BUREAU")
bureau <- makedummies(bureau,basal_level = TRUE) %>% # make dummy
  select(-SK_ID_BUREAU) %>%                          # not need
  SummarizeFunc(group = "SK_ID_CURR")                # summarize
```

## output for csv

- Save the csv files in the csv_imp folder

```{r eval=TRUE}
write_csv(bureau,"~/Desktop/Home_Credit_Kaggle/csv_imp/bureau_imp.csv")
```
