==> Rmd/1_Preprocess_pre_app.Rmd <==
---
title: "Preprocess_pre_app"
author: "kotsubo takuto"
output: 
    html_document:
      md_extensions: -ascii_identifiers
      toc: true
      toc_depth: 3
---

## Data description

- previous_applications.csv (37 variables)
    - Key variable: SK_ID_CURR, SK_ID_PREV
    - About: サンプルで融資を受けているお客様のホームクレジットローンのすべての以前のアプリケーション

# Setting{.tabset .tabset-fade .tabset-pills}

## knitr option

```{r reset, include=FALSE}
# 初期化
rm(list = ls())
```

```{r set up, message=FALSE}
# set directory
setwd("~/Desktop/Home_Credit_Kaggle/") 
# max.print 
options(max.print="200", digits=5)
# Global options
library(knitr)
opts_chunk$set(echo=TRUE,
               cache = TRUE,
	             prompt=FALSE,
               tidy=TRUE,
               comment=NA,
               message=FALSE,
               warning=FALSE)
opts_knit$set(width=75)
```

## Library package

- 必要なパッケージを適宜追加

```{r package, message=FALSE}
library(tidyverse)
library(readr) # for csv
library(summarytools) # summary easily for EDA
library(GGally) # ggpairs
library(skimr) 
library(missForest)
library(imputeMissings)
library(mi)
library(doParallel)
library(qgraph)
library(psych)
library(fastcluster)
```

## Load funciton

```{r cache=FALSE}
source('~/Desktop/Home_Credit_Kaggle/script/function.R') # for preprocessing
source('~/Desktop/Home_Credit_Kaggle/script/makedummies.R') # for preprocessing
```

## read csv

- Missing value
na = c("XNA","NA","","NaN","?"): 以下の設定で naを計上する.

- convert charcter type into factor type

```{r}
# クレジット利用履歴 (by Credit Bureau)
previous_app <- read_csv("~/Desktop/Home_Credit_Kaggle/csv/previous_application.csv",na = c("XNA","NA","","NaN","?")) %>% 
  mutate_if(is.character, funs(factor(.)))
```

# Preprocess

- Check for imputation values
- Describe the details on the explanation sheet
- Add `_imp` to the supplemented column
- Add `_add` to the added column

## check value

- 外れ値をNAとする

```{r}
previous_app <- previous_app %>% 
  mutate(DAYS_FIRST_DRAWING = ifelse(DAYS_FIRST_DRAWING == 365243, NA, DAYS_FIRST_DRAWING),
         DAYS_FIRST_DUE = ifelse(DAYS_FIRST_DUE == 365243, NA, DAYS_FIRST_DUE),
         DAYS_LAST_DUE_1ST_VERSION = ifelse(DAYS_LAST_DUE_1ST_VERSION == 365243, NA, DAYS_LAST_DUE_1ST_VERSION),
         DAYS_LAST_DUE = ifelse(DAYS_LAST_DUE == 365243, NA, DAYS_LAST_DUE),
         DAYS_TERMINATION = ifelse(DAYS_TERMINATION == 365243, NA, DAYS_TERMINATION))
```


## impute missing values{.tabset .tabset-fade .tabset-pills}

- 検証内容
    - データ特性に従って欠損値を推定
    - 多重代入法などのアルゴリムに従って, 欠損値を推定
    
### NAME_TYPE_SUITE

```{r eval=TRUE}
previous_app <- previous_app %>% 
  mutate(NAME_TYPE_SUITE_imp = replace(NAME_TYPE_SUITE, is.na(NAME_TYPE_SUITE),"Unaccompanied")) 
```

### AMT_

- AMT_ANNUITY
- AMT_APPLICATION (欠損なし)
- AMT_CREDIT
- AMT_DOWN_PAYMENT
- AMT_GOODS_PRICE

```{r eval=FALSE}
previous_app %>% 
  select_if(grepl("AMT",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

previous_app %>% 
  select_if(grepl("AMT",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)
```

```{r}
# data
data1 <- previous_app %>% select_if(grepl("AMT",names(.)))
# str_c(a,collapse = "', '")
patterns <- c("AMT_ANNUITY","AMT_CREDIT","AMT_DOWN_PAYMENT","AMT_GOODS_PRICE")
# 欠損値補完
tmp1 <- ImputeMissingValueMI(data1,patterns)
# データ結合
previous_app <- cbind(previous_app,tmp1)
```

### DAYS

- DAYS_DECISION
- DAYS_FIRST_DRAWING
- DAYS_FIRST_DUE
- DAYS_LAST_DUE_1ST_VERSION
- DAYS_LAST_DUE
- DAYS_TERMINATION

```{r eval=FALSE}
previous_app %>% 
  select_if(grepl("DAYS",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

previous_app %>% 
  select_if(grepl("DAYS",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)
```

```{r}
# data
data1 <- previous_app %>% select_if(grepl("DAYS",names(.)))
# str_c(a,collapse = "', '")
patterns <- c("DAYS_FIRST_DRAWING","DAYS_FIRST_DUE","DAYS_LAST_DUE_1ST_VERSION","DAYS_LAST_DUE","DAYS_TERMINATION")
# 欠損値補完
tmp1 <- ImputeMissingValueMI(data1,patterns)
# データ結合
previous_app <- cbind(previous_app,tmp1)
```

### RATE

- 特に関係ない

```{r eval=FALSE}
previous_app %>% 
  select_if(grepl("RATE",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  qgraph(edge.labels=T)

previous_app %>% 
  select_if(grepl("RATE",names(.))) %>% 
  na.omit() %>%
  cor() %>% 
  cor.plot(numbers = TRUE)
```

### NAME

```{r eval=FALSE}
previous_app %>% 
  select(NAME_PORTFOLIO,NAME_GOODS_CATEGORY) %>% 
  table(exclude = NULL)

previous_app %>% 
  select(NAME_CLIENT_TYPE,NAME_GOODS_CATEGORY) %>% 
  table(exclude = NULL)

previous_app %>% 
  select(NAME_SELLER_INDUSTRY,NAME_GOODS_CATEGORY) %>% 
  table(exclude = NULL)
```

## add new column{.tabset .tabset-fade .tabset-pills}

- 新しい特徴量を作成する

### AMT

```{r}
# data
data1 <- previous_app %>% select("AMT_APPLICATION","AMT_ANNUITY_imp","AMT_CREDIT_imp","AMT_DOWN_PAYMENT_imp","AMT_GOODS_PRICE_imp")
# str_c(a,collapse = "', '")
patterns <- c("AMT_TSNE_1","AMT_TSNE_2")
# 欠損値補完
tmp1 <- ImputeMissingValueTSNE(data1,patterns)
# データ結合
previous_app <- cbind(previous_app,tmp1)
```

```{r}
previous_app <- previous_app %>% 
  mutate(APP_CREDIT_PERC_add = AMT_APPLICATION / AMT_CREDIT,
         CREDIT_ANNUITY_PERC_add = AMT_CREDIT / AMT_ANNUITY,
         CREDIT_GOODS_PERC_add = AMT_CREDIT / AMT_GOODS_PRICE)
```

## Summarize

- *NAがInf or - Infになることに注意する*(max,min 計算時)

```{r}
previous_app <- makedummies(previous_app,basal_level = TRUE) %>% # make dummy
  select(-SK_ID_PREV) %>% 
  SummarizeFunc(group = "SK_ID_CURR") 
```

## output for csv

- Save the csv files in the csv_imp folder

```{r eval=TRUE}
write_csv(previous_app,"~/Desktop/Home_Credit_Kaggle/csv_imp/previous_application_imp.csv")
```
